BuildContext provides shouldExecuteInSeparateProcess and DefaultJavaOptions
Summary: Right now, we are having to pass these values into every `BuildableWithExternalAction`, which is really cumbersome and makes every method/constructor take two extra params. It's reasonable for these to be part of the `BuildContext` instead which will mean that the Steps/Buildables themselves don't have to worry about it.

Reviewed By: mykola-semko

fbshipit-source-id: 67d4e40967771598d691c7485d3553e52195d402