Use only `ExternalLogHandler` for external processes (buildables v2 and javacd)
Summary: JavaCD uses a standard buck's logger that has some default logger handlers associated with it (ex. console handler). This predefined loggers produced output to the console (std output, std error) that are consumed by buck process as a result of a launched process. We not really need to output logger events into a console as we register `ExternalLogHandler` that would communicate log events back to buck.

Reviewed By: irenewchen

fbshipit-source-id: cd232405977d0b924e88851faa8d3f22ac724f05