R.java files should be a dep of testsjar, not included in it
Summary:
Right now, the "test classes" of a robolectric test include all the R classes. That is because the R.java files are included as sources of testsjar. Instead, they should be built in a library that is then a dep of testsjar.

For example, we no longer see all the R classes referenced in this test description: https://www.internalfb.com/intern/paste/P168250416

Reviewed By: asp2insp

fbshipit-source-id: df703fdd6e44336f279235ed39c5e9309d911892