Monitor if WT process is still alive while waiting for a ResultEvent
Summary:
The issue:
Sequence:
1. Launch WT
1. Execute command that require wait for ResultEvent from launched WT
1. Repeat 2. if multiple commands
1. Shutdown WT

But if WT failed the current code is stuck in step 2 with waiting for a ResultEvent!

Solution:
During the launching a process simultaneously start a thread that would wait for a process exit code.
Then if WT failed, our  wait for launcher process task would receive non zero exit code of the terminated process and then cancel waiting for a ResultEvent that would unblock client.

Reviewed By: bobyangyf

fbshipit-source-id: c24c08c5dc78d25caeeec162dbde9c240403cfa0