Fix bug using headers generated from `cxx_genrule`
Summary:
If a header is generated by a `cxx_genrule`, even though it's listed
in the apparently platform-agnostic `exported_headers` parameter,
it requires propagating the C++ platform, and so is actually treated
as a platform exported header.

This fixes a bug which prevented the platform exported headers to
be generated in this case.

Reviewed By: mzlee

fbshipit-source-id: 4e6759296e6cbc460c8b2ea2c043a495988dad43