Increase debugger poll timeout
Summary:
Increased the poll timeout to resolve debugger attaching issue.

There is a possible ongoing race condition when running tests with the --debug flag where the BUCK will try to attach the debugger too early and in some cases, mainly larger targets, it will fail.

IntelliJ implementation has polling mechanism where it will retry the creation and attaching the debugger, increasing the pooling time to 30s is giving us enough time for BUCK to execute.

Once the polling timeout is reached to 0 and the debugger was not successfully attached the UX is bad since there is no feedback to the end user and the IJ/CLI are stuck in
`Debugging. Suspending JVM. Connect a JDWP debugger to port 5005 to proceed.`

This should not have a negative impact on smaller targets where the debugger is attached almost instantly.