Windows Named Pipes: handle end of the pipe
Summary:
Add special checks to Input and Output streams implementations for Windows Named Pipes that allow streams to handle the end of the named pipe correctly.

When I ran the code on windows with javacd I have spotted exceptions with the WinError.ERROR_BROKEN_PIPE error code. And this is a valid case that we should handle. After this change everything works.

Reviewed By: bobyangyf

fbshipit-source-id: e0b3060b5f534a75b242e5f346eeda0fbdf8ae2d