DefaultWorkerToolExecutor: open command's stream from named pipe only after external process is launched
Summary:
Named pipes implementation is different for windows and posix. On windows an operation of opening the stream from named pipe is blocking as server waits for a clients to connect. That is why windows blocks `namedPipeWriter.getOutputStream()` before external process that would connect to a pipe is launched.

This diff is changing the order of initializing fields for `DefaultWorkerToolExecutor`. Firstly the external process is launched and only then output stream is trying to be opened.

Reviewed By: stepancheg

fbshipit-source-id: 01048181e5a60754092286a005847c2b67edf34b