Don't pull in C++ libs for irrelevant platforms
Summary:
Currently, the native library gathering done for APKs doesn't respect
`platform_deps`, and so it can pull in dependencies that were never
intended for the platforms packaged by the APK.

This diff adds basic supported to filter out C++ deps that aren't relevant
to the APK C++ platforms.

The integration test below gives an example of the scenario this fixes.

Reviewed By: IanChilds

fbshipit-source-id: d197a8ebb95ace5c513bacb5ebdd14e62ec89828