Fix bug with conditional relinking and apple_toolchain
Summary:
### Context + Bug

Conditional Relinking needs to compute the file hashes of a set of `SourcePath` which are stored as keys in a map. It's actually possible for two *different* `SourcePath`s to map to the same path.

The issue arises because we put the mapped paths of the `SourcePath`s as keys in a dictionary. So, we ended up generating duplicate keys in the map builder.

### The Fix

Use a set to ensure we do not process the same path twice and thus avoid having duplicate keys.

Reviewed By: blackm00n

fbshipit-source-id: 24685b60656a384fc067a20b566c1628c9116b72