Updating shared module names to be concatenated string of all modules
Summary:
for better clarity on what shared modules are representing, updating naming convention of shared modules to concatenate underlying modules. this also helps us with any potential prefs stored for shared modules. eg if shared0 is a combination of foo and bar, but becomes a dep for foo, bar, baz tomorrow, all the previous prefs are not cleared.

i verified that google play does not seem to have a restriction (tested up to module with 180 char length) on module name length. the only restriction is from having a file name longer than 255 chars. eg: AndroidManifest__modularityapp_bundle_debug#a1234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890,manifest_merge__.xml which is 256 characters throws an exception at build time.

i would imagine that if a shared module length is this long, we can revert back to original naming convention (shared0, shared1,...). have not yet investigated how we would be able to detect that a file name would be too long at this stage. we would just have a build break in the meantime.

Reviewed By: longinoa

fbshipit-source-id: 970647f1142015c020b7ed77def360fd06d5f3e2