Opt-in support for using rule-specific host platform switching
Summary:
This adds optional support which allows using the rule's configured
"default target platform" as the host platform when performing host
platform switching (e.g. for the `$(exe ...)` macro).

Currently, the macros used in fbcode and fbsource used somewhat
complicated logic to apply proper target configurations for host builds
which are difficult to replicate via the host platform detector spec
(which currently just picks a global host platform configuration based
on OS).

For example, in fbcode, a rule should us either platform007 or
platform009 for its host platform depending on the project base path
(and whether it's opted in to the newer platform).  Another example
is that, on linux, we may want to use the fbcode platforms or the linux
platforms depending on the project.

In its current form, this diff just uses the `default_target_platform`
setting on rules, but it may make sense to also add a
`default_host_platform` parameter.

Reviewed By: mzlee

fbshipit-source-id: e711e60d02715fde3353c476482c0c798f3b71c9