Fix serialization of rust rules
Summary:
Another implementation of D25495175. Last time it broke serialization of rust rules for RE. This should fix it.

Note that path stored in string is not very multiplatform friendly but actually rustc should accept it on all platforms including Windows. Unfortunately, `Path` objects are forbidden in `Arg`. For C++ we also serialize `StringArg` with include path, so it should be fine.

Reviewed By: jsgf

fbshipit-source-id: 7f030baa1b99077f519bbf315d63f5472b6b49e4