javacd: borrow WT instances from the pool with timeout
Summary:
Switch to use a wait with timeout to prevent buck of stuck-ing forever when waiting for an available javacd worker tool form a pool.

Put a 30 minutes timeout to start with (

NOTE: May be it could be a valid case for android to build with 4 workers and wait for a job completion for more than 30 minutes. But let me adjust these values later in case it would be a real problem.

Reviewed By: irenewchen

fbshipit-source-id: ffbba89dbbed2a386f4deca40f55aa6d60ab55d7