Fix overwriting of last line of Starlark error
Summary:
We created multiple instances of `DirtyPrintStreamDecorator` for stdout or stderr. So when Starlark error was printed, `dirty` flag was set in a different instance of `DirtyPrintStreamDecorator` than one used in super console. So super console did not understand it should stop printing updates.

I'm not sure the error is specific to Starlark: the error is not reproduced on every invocation, it is possible that with Starlark we get this error more often.

This diff prevents creation of multiple instances of `DirtyPrintStreamDecorator` for each of stdout or stderr.

Reviewed By: aniketmathur

fbshipit-source-id: 9fc26cdaf8eef311d8d7bfc495563ce0c601c308