Remove hasFocusedTargets from DarwinLinker
Summary:
We're seeing increased cache misses in fbios and IG builds with the recent Buck bump carrying D26640944. With `hasFocusedTargets` as part of the rule key, we're seeing rule key differs on lower level xplat binaries such as `//xplat/msys/tools/sync/minizip:miniunz#binary`, which in turn triggers rule key differs on thousands of build artifacts.

To address this problem, this diff removes the use of `hasFocusedTargets` from `DarwinLinker`. We don't actually need this field and removing it would solve this issue.

Reviewed By: milend

fbshipit-source-id: 9aa42270843f07976544c43d6cc40e01f8c28010