Remove unused StarlarkThread parameters
Summary:
This makes calls slightly cheaper, and code cleaner.

It is faster, because there's another optimization which allows reusing the parameter array. E. g. for call `len(x)` Starlark internally creates an array `{x}` at the call site. And in the function implementation with `StarlarkThread` parameter it need to reallocate it to fit `{x, thread}`, and without thread parameter, it can reuse the array.

Reviewed By: mykola-semko

fbshipit-source-id: 615742da69eee9d74e017d8127b7f34cc968a388