Convert ScriptTest to junit parameterized test
Summary:
Include `ScriptTest` into a set of tests run by Buck to prevent regressions.

Before this commit `ScriptTest` was a runnable class with `main` function, but not junit test. Bazel executes it as a separate target, and we did not execute it before this diff.

Upstream PR: https://github.com/bazelbuild/bazel/pull/13200

Reviewed By: mykola-semko

fbshipit-source-id: 17ff0c322353276419b1a28637e09ff2457da08e