Recommit of "Updating shared module names to be concatenated string of all modules" - fixed with char limit check and negative hash check
Summary:
pushing this because and adding check to make sure hash value is positive (& 0x7FFFFFFF) to make sure shared modules dont have a dash in them.

 ---

pushing this out yet again after breaking ig since their shared module size exceeds 50 chars. only additional change is adding in a check in apkmodulegraph to limit the size of module names to be under 50 chars. tests are updated to used s_1018759649 as a sharedmodule name.

 ---

pushing out this change again. this commit was backed out before because it failed a test in fbandroid that hardcoded the expect shared module now. previously failing test are now ignored and will be updated after this lands.

 ---
Original commit changeset: f233df6ed50a

for better clarity on what shared modules are representing, updating naming convention of shared modules to concatenate underlying modules. this also helps us with any potential prefs stored for shared modules. eg if shared0 is a combination of foo and bar, but becomes a dep for foo, bar, baz tomorrow, all the previous prefs are not cleared.
i verified that google play does not seem to have a restriction (tested up to module with 180 char length) on module name length. the only restriction is from having a file name longer than 255 chars. eg: AndroidManifestmodularityapp_bundle_debug#a1234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890,manifest_merge.xml which is 256 characters throws an exception at build time.
i would imagine that if a shared module length is this long, we can revert back to original naming convention (shared0, shared1,...). have not yet investigated how we would be able to detect that a file name would be too long at this stage. we would just have a build break in the meantime.

Reviewed By: pcerl017

fbshipit-source-id: 4203e147f4a7c561df44773195323dfafed5cec2