CopyFile modern rule
Summary:
CopyFile is very similar to CopyTree, but removes the usage of the intermediate
`copy_tree` directory, and only works for one single file, so later rules can
use its output directly as input.

The ProjectFilesystem APIs are avoided since they will execute the moment the
steps are recovered, which might mean that the files are not already in place to
be copied.

Reviewed By: milend

fbshipit-source-id: 004c9c4097ef524fd402557f1f87f34d404f0446