Replace usage of WriteFile for CopyFile. Use SourcePath for stub binary.
Summary:
Use the new CopyFile in the places in the Apple rules where a WriteFile was
being used, to correctly record the dependency on the origin file as
a SourceFile.

Instead of using `Path` use `SourcePath` for the WatchKit stub binary. Use `PathSourcePath` for the Xcode toolchain, and use the optional value of `stub_binary` when configuring a toolchain.

Reviewed By: milend

fbshipit-source-id: 9f8f904f7f8810a612aac22f6bc9d5e33a0624b4