Support UUID scrubbing of 2GiB+ Mach-O files
Summary:
Adds support to be able to scrub 2GiB+ Mach-O files. In this particular diff, the hashing is amended to support larger file sizes.

The underlying Java APIs use a 32bit signed int for the memory map size, so there's a limit of 2GiB. The way to work around is to hash multiple chunks, each 2GiB of size. That works because the memory map offset is a 64bit signed int.

Reviewed By: blackm00n

fbshipit-source-id: 1abbbc10ef2e46baf7ee4cf53f1ac0565ba14ac6