Correctly propagate linker normalization flags
Summary:
`DefaultLinkerProvider` was passing a default `false` parameter for the linker normalization args flag. We need to be using linker normalisation args as ARM64 executables running on a Mac are adhoc signed by default which breaks the OSO relativizer in Buck.

As we were passing `false`, there were a few places which ended up passing the wrong value and unable to build ARM64 Mac executables.

NOTE: All places which hardcode passing of `true` for the linker normalization flags are only used in tests.

Reviewed By: blackm00n

fbshipit-source-id: f7856d45bcee217e99309af386fb6c165faff17a