Correctly handle dSYM data with cross-cell builds and link path normalization
Summary:
# Context

Mach-O executables store *absolute* paths to the object files when linked. In order to produce hermetic executables, those absolute paths have to be relativised. There are two ways to do that:

- The `OsoSymbolsContentsScrubber` which lives in Buck.
- Passing `-oso_prefix` to the ld64 linker (controlled via the `cxx.link_path_normalization_args_enabled`).

Using `OsoSymbolsContentsScrubber` is problematic for two reasons:

- It has a limit of 2GiB executables due to memory mapping APIs in Java using signed 32bit integer
- It does not play well when building for ARM64 for Mac because, by default, the linker will adhoc sign the binaries (as all binaries need to be code signing on ARM64 macOS).
  - While this linker behavior can be disabled by passing `-no_adhoc_codesign`, it will cause further issues as the binaries won't run on ARM64 Macs without further manual intervention.

So, the overall goal would be to just deprecate `OsoSymbolsContentsScrubber` and globally rollout `cxx.link_path_normalization_args_enabled`.

# The Bug

After passing a path prefix to the linker (`-oso_prefix`), all paths in the executable will relative. When `dsymutil` processes an executable, it needs to be able to find all those object files to extract the DWARF data out of. It can do so in two ways:

- Set the cwd directory to the prefix that was stripped. This is a bit fragile.
- Pass `--oso-prepend-path=` parameter to `dsymutil`, so that it can find all object files.

Previously, we were relying on the cwd just being correct without any enforcement. In the case of cross-cell builds, this turned out to be incorrect.

# The Fix

`HasAppleDebugSymbolDeps` is extended to expose the OSO prefix which is then consumed by `AppleDsym` to correctly pass the path.

Reviewed By: blackm00n

fbshipit-source-id: aa743e54cf161d40492f0c6cfcdf122870776f6c