Assert watchman has clock-sync-timeout capability
Summary:
According to xavierd, everywhere in Facebook we use the same version of watchman, one month old in the worst case, and when we don't, this is an infrastructure error which needs to be fixed.

Capability detection was somewhat needed when our Buck was full open source, but this branch is forked which will likely never be properly open sourced again.

So, instead of implementing different Buck behavior depending on Watchman capability, let's just assert that Watchman is fresh.

This way we:
* assert watchman is indeed fresh everywhere, and our slow builds are not caused by old watchman
* simplify Buck logic
* prevent ourselves from possibly incorrect behavior, e. g. if there's a typo in capability name, we should crash instead of working without the capability.

Reviewed By: bobyangyf

fbshipit-source-id: d0e02637ee75625c4cab358c8d64764ed96ccba9