watch-project only once
Summary:
Problem with current implementation is that it's possible that first watch-project takes like 11.9s of 12s timeout, so following watch-project would have to be completed in 0.1s, which it will likely not, and we will work without watchman.

To fix it we need to either increase timeout of second watch-project, or just remove this old code.

This TODO was added in D3939383 5 years ago by mzlee and wez commented that it will be fixed soon. Hope it was fixed.

Reviewed By: mzlee, wez

fbshipit-source-id: 9d93ccbb7ad0bb8a6b6e2e102a4d97684928e2fd