prebuilt_cxx_library now supports pre-built DLLs with no associated static import libraries on Windows
Summary: On Windows this enables prebuilt DLLs, for which there are no associated static import libraries, to be added as dependencies of other targets and to be added to the binary with shared library symlink tree properly. This is useful for prebuilt libraries that have only a single static import library but come with multiple DLLs and plugin-style systems that load DLLs at runtime.

Reviewed By: akrieger

fbshipit-source-id: b080c1cd334c578ac0d899fa21bc6736b0777250