Disable sync cookies in watchman globber
Summary:
Since D27179706, we ensure that clock request was successful when creating watchman.

Clock request performs sync.

We create watchman in `Main`, in the beginning of each command. So we always perform the sync in main.

So there's no need to perform sync again when using watchman globber. So glob (in Starlark in particular, but in other places which use `WatchmanGlobber` too) should be faster, especially if watchman suddenly decided to recrawl.

Reviewed By: mzlee

fbshipit-source-id: aa1cb6d9955a7b636374662b6ade031e125d0576