Convert StarlarkValue to class [REVIEW HERE]
Summary:
Interfaces are really slow compared to classes.

This is usually not an issue when at given call site practically there's only one class, but this is not the case for Starlark: there are too many different `StarlarkValue` classes, and JVM gets confused.

So convert `StarlarkValue` to class.

This makes code slightly less maintainable and more diverged from upstream, but that's price I'm willing to pay.

Most of the changes in this stack are mechanical.

This simple change leads to surprisingly high 6% speedup:

```
A: n=20 mean=76.386 std=1.782 se=0.408 min=73.644 med=76.229
B: n=20 mean=71.405 std=1.216 se=0.278 min=69.309 med=71.218
B/A: 0.935 0.922..0.948 (95% conf)
```

Reviewed By: bobyangyf

fbshipit-source-id: 42682c0d214ce9fd0fac1c5e7ba3e3c6bdd29971