Support for macOS platform target triples
Summary:
Previously, we were not setting the `--target` parameter when compiling for macOS and Catalyst targets on Apple platforms. That meant we were always compiling for the host architecture, so cross-compilation was not possible.

In this diff, we add support for translating the macOS targets to their `--target` equivalents in Rust, so we can cross-compile from Intel Macs to ARM64 Macs.

Reviewed By: blackm00n

fbshipit-source-id: cc2d32635a9b2500b78b6a596154c1ebc3db8a72