Remove rapid install option
Summary:
This option was added in D6140297. However, it was never turned on, and per discussion here https://fb.workplace.com/groups/442333009148653/permalink/4607233239325255/?comment_id=4633179433397302&reply_comment_id=5395728370475734 it is not needed since multi-install (D6304020) does the useful parts of this anyway.

Therefore removing to simplify the code. Will simplify the protocol and remove more parts in the next diff.

Reviewed By: christolliday

fbshipit-source-id: 9009d4fa8b7c8c76d0f218acc59ba50bbafeee81