Optimize linking for BuiltinFunction
Summary:
Builtin functions are usually called with positional arguments only (e. g. `list(x)`), and full linking for them won't help much.

However, default linking implementation of linking in the previous diff creates a linked function which delegates to `fastcall`, which is inefficient for `BuiltinFunction`.

This diff introduces `StarlarkCallable.linkAndCall` API which is glued link+call operation, and this operation is implemented natively for `BuiltinCallable` instead of previous `fastcall`. `linkAndCall` operation is used with `CALL` instruction (when link target is not known). So this operation avoid overhead of linked function allocation for builtin functions.

Full explanations and results are in the top diff D27028973.

Reviewed By: andrewjcg

fbshipit-source-id: be8ac0f99705d1cfa555cbc9197c35d99729d553