PooledWatchmanClient
Summary:
`PooledWatchmanClient` implements `WatchmanClient` interface, but holds a pool of `WatchmanClient`.

The goal of having pooled watchman client:
* avoid creating connection for each operation (we do that in certain parts of Buck; connection open is cheap but not free)
* avoid creating more connections than needed

The main motivation for the diff is to switch starlark parser back to single instance (instead of instance per thread) to save memory and to be able to use pooled watchman client there.

Pool is a really simple `ConcurrentLinkedQueue`.

We have `ResourcePool` class, but it is heavily overengineered, so this diff implements pooling (5 lines of code) from scratch.

This diff only implements the pool, but does not enable it anywhere.

Following diffs switch code to pooled watchman client one site per diff, to be able to bisect if something goes wrong.

Reviewed By: miaoyipu

fbshipit-source-id: 666a46402db464ad561a5b1ab88886f85c360076