Drop AST after .bzl evaluation
Summary:
AST takes quite a lot of memory. Drop it when it is no longer needed: right after bzl is evaluated.

`extensionCache` field is now a map to either `Program` (parsed) or `ExtensionData` (evaluated).

Additionally, we no longer cache AST for `BUCK` files. It is not really needed because we create parser per-build, and we evaluate each `BUCK` file at most once.

Reviewed By: mykola-semko

fbshipit-source-id: 6de4ba66e71373dd560b103fec1af771ed908210