Fix an exception during highlight process
Summary: When there is unresolved element in the file, e.g. an android resource file - R.java class, it will not have the corresponding virtual file, this leads to a crash in the middle of highlighting process, which causing to not highlight/not highlight correctly some parts of the code.

Reviewed By: cute-jumper

fbshipit-source-id: 3280491caa9d08ab5029bc2e44402e4e854b1268