Fix conditional relinking when used with apple_toolchain.
Summary:
When writing conditional relinking info there's a step where we deduplicate the source path. The deduplication needs to happen in the conditional relinking check as well, otherwise the number of source paths and the number of conditional relinking files won't match when there are duplicated source paths.

This should fix this issue where CR's not working on apple toolchains, including with pika and in on-demand.

Reviewed By: milend

fbshipit-source-id: d3d102ffd79606ba5cdc7a8d19c12b806c570b3c