Remove DX option from DxStep
Summary:
`dx` is old and now we use `d8` everywhere we can simplify the code by just retaining that code path.

As ever with Buck cleanup, a lot of the file changes are just removing unused vars/params. Interesting stuff is in `DxStep`. I will rename that file in a future diff.

Reviewed By: christolliday

fbshipit-source-id: 7e031b24593c83901460a19b48e6cf034f97b84a