BcVisitor -- bytecode visitor
Summary:
`BcVisitor` class is a visitor for bytecode.

Useful to perform static analysis on bytecode, which is sometimes easier and faster than performing static analysis on AST.

Easier because everything is resolved (e. g. constants).

Faster because no hashmap lookup by string is performed in bytecode unlike when working with AST, and because flat bytecode iteration is faster than visiting the tree.

Default implementation of `BcVisitor` just performs some assertions which are enabled when assertions enabled (off by default).

`BcVisitor` is used in the following diff D27559923 to implement function call caching.

Reviewed By: andrewjcg

fbshipit-source-id: 091c26f1f587c70aeae8b70c1c62b5162a96b1da