Optimize const str.format and other getattr calls
Summary:
Consider this program:

```
foo = {"a": "b"}

def f():
  return foo.get("a")
```

Before this diff it was compiled to:

```
$1 = getattr(foo, "get")
$2 = call(foo, "1")
return $2
```

now it is compiled to

```
$2 = call(foo_get, "1")
return $2
```

Note this optimization is safe even if captured variable is mutable.

This change assumes that all types `getattr` produces the same instance during object lifetime. Which is the case for all types.

Previously only `struct` `getattr` was optimized, this diff extends it to any types.

This diff also contains some getattr-related code cleanup.

Reviewed By: andrewjcg

fbshipit-source-id: d4eb0c4f55c0bfab65e006f8310f53c6ffc40593