Add option to not fail dSYM verification
Summary:
If `dwarfdump` is missing, we cannot extract the warnings from the dSYM and do verification. Previously, the build would fail if dSYM verification was requested but `dwarfdump` was missing.

After this diff, the behaviour can be configured, so that the verification gets skipped if `dwarfdump` is missing by setting `apple.dwarfdump_fails_dsym_verification=false`.

Reviewed By: blackm00n

fbshipit-source-id: 57ae9af5adeb38187c65d9ef7c4d7963050e57bf