Add ability for PreDexGroups to also record which classes are needed in the primary dex
Summary:
We need this in order to only add the relevant classes to the primary dex, and every class in the jar.

We change the `PreDexedFilesSorter` to also return the relevant classes for each jar. This information will be used in a future diff by the `PreDexedSplitDexMerge` rule.

This is behind a config flag so that we can easily turn this on and off, since this could manifest as a runtime error.

Reviewed By: christolliday

fbshipit-source-id: f50c183ac5ba6c3d4e522f72f122cb04170a4d68