Update run_inplace.py.in to allow passing initial commands to pdb
Summary:
Add support for passing initial commands to pdb when running a binary
under a debugger. We cannot use the `-c` argument to pass commands to pdb,
because the startup wrapper is a script.

Read initial debugger commands from the `PYTHONPDBINITIALCOMMANDS` environment
variable. Commands are separated by a `|` character, but passed to pdb verbatim
otherwise. The `pdb.rcLines` call works on Python 2.x and 3.x.

Reviewed By: loganwendholt

fbshipit-source-id: da3e9b8eb515ff915ae28cf5339cf7676ebf2bee