LoadedModule
Summary:
Before this commit Starlark had `Module` class which was used as both:
* evaluation context
* returned value for `load` callback, a mapping from global name to global value

`Module` object contains not just globals, but also universe (e. g. `True` and `None`, which are not visible to load).

This diff introduces `LoadedModule` interface which is used in `load` callback.

This make code more clear (for example, it allows providing custom module implementation which is not based on evaluation `Module`).

This also makes following changes (split module and resolver module) possible.

Reviewed By: aniketmathur

fbshipit-source-id: cc7cadfab2b8b7c0055a38fdcd49a5c06515a74d