PreDexSplitDexMerge only puts specified classes into the primary dex
Summary:
Before now, `PreDexSplitDexMerge` was lazy and if a single class from a pre-dexed jar was needed in the primary dex, it would just put the whole thing into the primary dex. This means that the primary dex got easily bloated, and it was hard for a big library, such as the Kotlin standard library, to be used in the primary dex because it would drag in so many classes and exceed the field/method count.

Instead, we can ask `d8` to just include the classes we actually need in the primary dex. This does give us a bunch of extra work to do, because the classes that didn't make it into the primary dex are spat out in RAW secondary dex format. To make them fit the same format that we have for our other secondary dexes in the same application, we run them through the `SmartDexingStep` again with a canary class. This does result in some code that feels, ahem, not ideal because we are dynamically creating steps within steps and then running them. I don't know how common that is in Buck and if there is a way to avoid it.

Reviewed By: jiawei-lyu

fbshipit-source-id: 2181fd05ecdec01640192fb9048edbc44e9843e4