Get rid of StarlarkFunction.globalIndex
Summary:
Before this diff, there was two-level addressing of globals:
* resolver assigned index to each global
* module has own index for each globals
And these indices were not the same. Because of that `StarlarkFunction` had a mapping from one index to another index (`globalIndex` field).

This diff changes all this machinery to query module index right from resolver, so second index is not needed.

That was high level overview.

To achieve that, this diff makes following changes:
* rename `Resolver.Module` to `ResolverModule`
* make `ResolverModule` concrete class instead of interface
* `ResolverModule` now holds a mapping from global name to index
* `Module` no longer extends `Resolver.Module`, but instead contains a field of type `ResolverModule`
* `ResolverModule` now has `freeze` operation which prevent modification of index. `freeze` can be called after the resolution, and we do it now in Buck

Before this diff in Buck we created the same module object twice:
* first to parse the file
* then to evaluate the file

Now the `ResolverModule` is created only once for the parsing/resolving, and `Module` is created for the evaluation.

Note currently in Buck we can evaluate the same module multiple times (when it is requested concurrently). Probably we should stop doing that.

Essentially, after this diff `StarlarkFunction` class no longer has `globalIndex` field.

Reviewed By: aniketmathur

fbshipit-source-id: af459b7fb35bb6322051145fadae2f304e0f3440