unnecessary copy in augmented assignment to local variable
Summary:
Operation like

```
a += b
```

Previously was compiled to two instructions (simplified):

```
temp = a + b
a = temp
```

now when `a` is a local variable, it is compiled to

```
a = a + b
```

Some context: instruction output (like +) is always a slot (local variable or temporary). So for global variables (or cells) we need to assign to temporary variable first and then reassign it to global, but for local variables we can assign to local variables directly.

Reviewed By: aniketmathur

fbshipit-source-id: 0f32df9734a395e0adf63a6048804c3371e2e9cc