Unnecessary temp variable in for loop
Summary:
Before this iteration, `for vars in` statement or for expression was compiled to

```
temp = iterator.next();
vars = temp
```

As in previous diffs, temporary variable is needed for `vars` is global or tuple (e. g. `for (x, y) in`).

When `vars` is a local variable, we can assign iterator to it directly avoiding temporary variable.

Reviewed By: aniketmathur

fbshipit-source-id: 3426a8ad92acef3b0f9544fa6535170d5666873c