PLUS_LIST and PLUS_STRING
Summary:
The idea is to
* replace dynamic dispatching with static dispatching where possible
* avoid list allocation where possible

Before this diff binary operation for say `x + y` where `x` and `y` are string were performing these operations:

* switch on operator
* if plus, switch on type, check int, tuple, string, list, ...

So it is several branches even if we know at compile time that operator is plus and arguments are string or list.

This diff introduces two new opcodes and changes semantics of several others.

## PLUS_STRING

`PLUS_STRING` is equivalent to `PLUS`, but it is optimized for the case when both arguments are string. This opcode is used when either argument is known to be string.

## IN_LIST

New bytecode operand introduced — `IN_LIST`. It represent either:
* length-delimited list of slots
* an `Object[]` in object pool

`IN_LIST` is a list of values.

## PLUS_LIST

`PLUS_LIST` takes two input arguments (lhs and rhs) and output argument.

`rhs` argument is `IN_LIST`

When `PLUS_LIST` is invoked with `list` argument, new list object is not allocated.

## LIST, TUPLE, CALL*

These instructions were updated to use `IN_LIST` operand to store arguments.

Reviewed By: aniketmathur

fbshipit-source-id: 7b906f1745f38525b5aff056ffd763c5f1e213a5