Use correct index when creating canary classes
Summary:
Right now, the index that we use for the canary class does not always match the index we use for the filename, which is confusing and can lead to errors.

To make it match, in the case that we don't have a `groupIndex`, we use `dexStore.dexClassNameIndex(index)`, rather than just `index`.

This removes the need for the `+ 1`  that we were passing in `PreDexedFilesSorter`, which didn't always work depending upon which type of dex class we were creating.

`DexTestUtils` now always verifies the canary class name - both that it matches one of the patterns we expect, and that it matches the dex filename that it is used in.

Reviewed By: christolliday

fbshipit-source-id: e34cb29448420972072ac9383117860f9b4c4c86