Rename AndroidBinary to AndroidApk
Summary:
As part of splitting apart `AndroidBinaryBuildable` it was pointed out
that the naming convention was a bit bizarre.

Overall we are going to migrate all usages of
`android_binary` -> `android_apk` in our BUCK files. This is going to
be a several diff approach.
1) rename AndroidBinary -> AndroidApk (this diff)
2) copy AndroidBinaryDescription -> AndroidApkDescription
N) move other files (factories?)
N+1) migrate all usages of android_binary -> android_apk
N+2) remove AndroidBinaryDescription

Reviewed By: IanChilds

fbshipit-source-id: b9fcc0765551957aad25bcdc30f2332de4fb40e7