Enable annotation processing in Idea
Summary:
Following diffs start generating `.java` files for Starlark.

So in order to run tests in Idea, we need to enable annotation processing in Idea too.

I have not found a way to tell Idea to fetch annotation processor from sources, it seems to only understand annotation processors in jar files.

So this workaround is used:
* `third-party/java/bazel/jars-for-idea` contains prebuilt annotation preprocessor
* `third-party/java/bazel/jars-for-idea/gen.sh` generates these jars
* `//third-party/java/bazel:jars-for-idea-verify` test verifies that prebuilt jars are up to date

Reviewed By: aniketmathur

fbshipit-source-id: 9e06b3cc252fe6dd7f8218355983d0a8039c7d70