Better way to find the zstd library path in the agent APK
Summary:
Before, we'd find the ABIs, and then assume that the zstd library was located in `libs/<some ABI>`. But that didn't work as we could get ABIs like `arm64-v8a` and the library would be in `libs/arm64`.

Instead, we can just look for the library in the exopackage's native library directory. If we don't find it, then we don't do zstd compression, but if we do then we can use that in our java library path.

Reviewed By: jiawei-lyu

fbshipit-source-id: 8058ab0e2a47d123ebcb6337515d14f7727c1938