faster list iteration
Summary:
Most iterated objects are lists.

Implement a special case for iteration over lists and tuples: when iterated object is a list or tuple, instead of creating an iterator object and invoking `hasNext`, `next` on each iteration, fetch the underlying array of the collection and do the indexed iteration.

We save on:
* iterator object allocation
* virtual iterator calls
* range checks

Reviewed By: aniketmathur

fbshipit-source-id: 796b863aed555733e3069508649a8a19fb06a6c7