Share locks for different cell caches
Summary:
The idea is to have a build number stored in `DaemonicParserState`, and update in holding this lock. To achieve this we need that lock.

We could instead hold multiple locks when updating that build number, but since builds are mostly single cell, single lock instead of multiple won't make difference.

Reviewed By: mykola-semko

fbshipit-source-id: a30b9332406bf453083108281abf2b28bc3ce147