Support collecting annotation processing events from Java builds
Summary:
In order to instrument Annotation Processing for Java builds, we're using existing events that are sent to build the Chrome Trace files. These SimplePerfEvents are sent when an Annotation Processor is started and finished.

This diff adds code for catching these perf events, accumulate them inside the new AnnotationsProcessorsLogger and generate a full perf stats data object to be stored and later logged in Scuba.

Reviewed By: IanChilds

fbshipit-source-id: 04bb243100d1fdbf14e2d792a21aca66426c18ef