Use ForwardRelativePath instead of AbsPath in parse pipeline
Summary:
All operations accept cell as parameter, so cell path of `AbsPath` is redundant.

This is done mostly to make it clear that passed path belongs to the same cell as next parameter. No runtime assertions needed.

Additionally, we verify that paths are normalized (i. e. no dot-dot in them).

Note even when cache interfaces are parameterized with `ForwardRelativePath`, cell is still a part of cache key, as cell is passed as function parameters.

This is quite dangerous change, but important to better understand what's going on in `DaemonicParserState` and to make sure `DaemonicParserState` works with paths correctly.

Reviewed By: mykola-semko

fbshipit-source-id: a0522bd3074bb42cd4bedccf1624c3c484d6ac84