StarlarkFunctionLinkedError
Summary:
Introduce `StarlarkFunctionLinkedError` utility.

`StarlarkFunctionLinkedError` is a proper function invocation which is known to produce an error.

The idea is this: proper error handling in Starlark invocation is hard, and it is even harder in high-performance code: we need to optimize code for no-error case but still produce nice error messages for error case (for example, create a variable to store missing arguments and populate it).

So, now when linking or invocation detects an error, it falls back to `StarlarkFunctionLinkedError`, which does linking again from scratch, but optimized for readability and correctness, not for performance.

`StarlarkFunctionLinkedError` is used in both linking (when linking is known to produce error) or in `StarlarkFunctionLinked` when error can be detected only in runtime.

Reviewed By: mzlee

fbshipit-source-id: 2a22d4cffa47fa76cbdb77e4ffdb8a1e7482781b